+++ src/limplungs/shards/ShardsGameData.java
@@ -2,6 +2,6 @@ package limplungs.shards;
 
 public class ShardsGameData
 {
-	public static int POSITION_X = 0;
-	public static int POSITION_Y = 0;
+	public static double POSITION_X = 0;
+	public static double POSITION_Y = 0;
 }
